pyproject.toml
[project]
name = "pcd-scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "requests-cache>=1.3.1",
]
